[ADD] tja weet het niet meer

+++ src/Automatic/Automatic.java
@@ -3,7 +3,7 @@ package Automatic;
 public interface Automatic {
 
     String getName();
-    int getCost();
+    double getCost();
     double getMultiplication();
     int update();
 }
